Consultar, Eliminar, añadir y modificar hoteles

+++ just-travel-it/src/main/java/is2/justtravelit/controllers/HotelRestController.java
@@ -42,7 +42,7 @@ public class HotelRestController {
     }
 
     @GetMapping("/hotels/delete/{id}")
-    public ResponseEntity<HotelDTO> deleteHotel(@PathVariable Integer id){
+    public ResponseEntity<HotelDTO> deleteHotel(@PathVariable Long id){
         HotelDTO response = hotelService.deleteHotel(id);
         return new ResponseEntity<>(response, HttpStatus.ACCEPTED);
     }
